backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.4.0
Pillow
django-extensions==3.2.3
pandas==2.3.3
numpy==2.4.0
openpyxl==3.1.2
xlrd==2.0.1